-	Updated to work with iOS 13 and Swift 5.1. -	Changed the ML Model from VGG16 to ResNet50. -	Tested on iP11Pro with iOS 13.1.
