+++ SayIt/Base.lproj/Main.storyboard
@@ -36,6 +36,7 @@
                                 <fontDescription key="fontDescription" name="AmericanTypewriter" family="American Typewriter" pointSize="15"/>
                                 <state key="normal" title="Flash Off">
                                     <color key="titleColor" red="0.77048004748774501" green="0.20114730136846404" blue="0.11764705882352941" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+                                    <color key="titleShadowColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
                                 </state>
                                 <connections>
                                     <action selector="flashBtnWasPressed:" destination="BV1-FR-VrT" eventType="touchUpInside" id="d1y-CJ-PmJ"/>
