+++ SayIt/View/RoundedShadowView.swift
@@ -1,0 +1,21 @@
+//
+//  RoundedShadowView.swift
+//  SayIt
+//
+//  Created by Ahmed AlOtaibi on 11/25/17.
+//  Copyright © 2017 Ahmed AlOtaibi. All rights reserved.
+//
+
+import UIKit
+
+class RoundedShadowView: UIView {
+    
+    // Apply shadow to the upperLabel
+    override func awakeFromNib() {
+        self.layer.shadowColor = UIColor.orange.cgColor
+        self.layer.shadowRadius = 15
+        self.layer.shadowOpacity = 0.75
+        self.layer.cornerRadius = self.frame.height / 2
+    }
+
+}
